Adds support to restructuredText markup
CSS:
introduces a new css file with rules adapted to the way rst2html
tool parses the .rst markup to html. Mainly/currently only
applying the proper css rule for inline code elements which is in
rst land is a ".literal" class to a "tt" element.

JS:
for highlight.js to work on code blocks generated by rst2html, the
code blocks (wrapped in <pre class="literal-block code") has to be
further wrapped in <code></code>.

To do that, I introduced a '$.hugo' namespace that should hold
whatever hugo variables we need along to the js land.